Design get potential supply

+++ contracts/LpToken.sol
@@ -7,7 +7,7 @@ contract LpToken is IERC20 {
     using SafeMath for uint;
 
     uint8 public constant decimals = 18;
-    uint public totalSupply;
+    uint256 public totalSupply;
     mapping(address => uint) public balanceOf;
     mapping(address => mapping(address => uint)) public allowance;
